+++ packages/eslint-plugin-stories/src/utils/includesAll.ts
@@ -1,3 +1,0 @@
-export default function includesAll<T>(list: T[], requireds: T[]): boolean {
-  return requireds.every((required) => list.includes(required));
-}