+++ packages/story-utils/src/generateSnapshots.ts
@@ -1,53 +1,76 @@
+import type { Meta } from '@storybook/react';
+import { composeStories, composeStory } from '@storybook/testing-react';
 import { render, RenderResult } from '@testing-library/react';
-import { getStoriesFromStoryFileExports } from './getStories';
-import type { StoryData, StoryFileExports } from './getStories';
-import prepareStory from './prepareStory';
+import { createElement } from 'react';
 import wait from './wait';
 
-type TestOptions<Args> = {
+type TestOptions = {
   /**
-   * get the snapshot either synchronously, returning the ChildNode,
-   * or asynchronously returning a promise that resolves to the ChildNode.
+   * Get the element to snapshot, either synchronously or asynchronously.
    */
-  getSnapshot?: (
+  getElement?: (
     wrapper: RenderResult,
   ) => Promise<ChildNode | null> | ChildNode | null;
-  /**
-   * Override the args for the story. Useful for event handlers that may be missing.
-   */
-  argOverrides?: Partial<Args>;
 };
 
-function getDefaultSnapshot(wrapper: RenderResult) {
-  return wrapper.container.firstChild;
-}
+/**
+ * Type for the star import of a stories file. For example:
+ *
+ *   import * as stories from './foo.stories.tsx';
+ *
+ * Technically there will also be stories themselves on the import, but there doesn't seem to be a
+ * good way to type those.
+ */
+type StoriesImport = {
+  default: Meta;
+};
 
 /**
- * Runs snapshot tests on all stories imported with `import * as snapshotTestStoryFile from 'storypath.stories.tsx'`
- * @param storyFileExports the exports of a .stories.tsx file including the default export with additional context
- * @param options the options for the test to add necessary context specific to snapshot tests
+ * composeStories uses composeStory. There's no explicit type exported that we can use while
+ * iterating stories (as of v1.0.0), so instead directly use the return type of the function.
  */
-export default function generateSnapshots<
-  S extends StoryFileExports<Args>,
-  Args,
->(
-  storyFileExports: S,
-  { getSnapshot = getDefaultSnapshot, argOverrides }: TestOptions<Args> = {},
-): void {
-  const stories = getStoriesFromStoryFileExports(storyFileExports);
+type ComposedStory = ReturnType<typeof composeStory>;
 
-  for (const [storyName, story] of Object.entries<StoryData>(stories)) {
-    if (story.parameters?.snapshot?.skip) continue;
+/**
+ * Generate snapshot tests for all stories imported from a file.
+ *
+ * @example
+ *
+ * import {generateSnapshots} from '@chanzuckerberg/story-utils';
+ * import * as stories from './SomeComponent.stories.jsx';
+ *
+ * generateSnapshots(stories);
+ */
+export default function generateSnapshots(
+  storiesImport: StoriesImport,
+  { getElement = (wrapper) => wrapper.container.firstChild }: TestOptions = {},
+): void {
+  const stories = composeStories(storiesImport);
+
+  for (const [storyName, Story] of Object.entries<ComposedStory>(stories)) {
+    if (Story.parameters?.snapshot?.skip) continue;
 
     test(`${storyName} story renders snapshot`, async () => {
-      const view = render(prepareStory(story.storyFn, argOverrides));
+      const view = render(createElement(Story));
+
+      // @storybook/testing-react doesn't run play functions automatically (as of v1.0.0). So if
+      // one is present, run it before taking a snapshot.
+      if (Story.play) {
+        const storyContext = {
+          loaded: {},
+          abortSignal: new AbortController().signal,
+          canvasElement: document.createElement('div'),
+        };
+
+        await Story.play(storyContext);
+      }
 
       // When components that include Apollo's useQuery are rendered we need
       // to await an act that pushes the test to the end of the event loop.
       // https://trojanowski.dev/apollo-hooks-testing-without-act-warnings/
       await wait();
 
-      expect(await getSnapshot(view)).toMatchSnapshot();
+      expect(await getElement(view)).toMatchSnapshot();
     });
   }
 }