+++ packages/eslint-plugin-stories/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - [breaking] Output ES2019 instead of ES6 - [#19](https://github.com/chanzuckerberg/frontend-libs/pull/19)
 - [breaking] Output [ES modules](https://nodejs.org/api/esm.html) instead of CommonJS - [#19](https://github.com/chanzuckerberg/frontend-libs/pull/19)
+- [breaking] Removed the `no-story-decorators` rule, because @storybook/testing-react will find and use decorators - [#18](https://github.com/chanzuckerberg/frontend-libs/pull/18)
+- [breaking] Removed the `no-top-level-story-args` rule, because @storybook/testing-react will find and use args at any level - [#18](https://github.com/chanzuckerberg/frontend-libs/pull/18)
 
 ## 1.2.2 (2021-08-13)